Move client side decoration rendering to server side
The client is still drawing its decoration client side to avoid
scaled rendering which make it unreadable when using the client zoom.
The server is also drawing its decorations but they do not appear in
sent frame to the client to avoid the double decorations drawing.
This ensure that both the client and server have their decorations
independently of each other.

Task-Id: KDEND-108