Unify the quick window and derivated content grabbing
This is needed since the introduction of the OverlayItem in the
QuickInspector as we want to filter out the overlay from the grabbing.
In fact we want to filter it out is because the overlay would then be
rendered at the resolution of the scene render it quite unreadable
client wise using small/high zoom.
So both the client and the server draw the decoration but at their
desired resolution.

Task-Id: KDEND-108